adding parameterized tests, not casting ofly params as string
