+++ requirements.txt
@@ -1,6 +1,7 @@
 coverage==3.6
 mock==1.0.1
 nose==1.2.1
+nose-parameterized==0.2
 pep8==1.4.2
 pyflakes==0.6.1
 requests==1.1.0
